requirements.txt
openai
langchain
langchain-community
langchain-openai
langchain-core
openai-whisper
faiss-cpu
streamlit
playwright